requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
Django==5.0
django-cors-headers==4.6.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.15.1
fonttools==4.55.1
fsspec==2024.10.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
identify==2.5.36
idna==3.8
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jq==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==75.6.0
Shimmy==2.0.0
six==1.16.0
sqlparse==0.4.4
stable_baselines3==2.4.0
stack-data==0.6.3
stripe==10.9.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
virtualenv==20.26.2
wcwidth==0.2.13
